feat: 모집 활성화 날짜 포맷팅 추가 (#10)
