+++ src/app/admin/(with-sidebar)/results/_components/mail-manage/ManageResultMail.tsx
@@ -2,7 +2,7 @@
 
 import {useState} from 'react';
 import {ManageMail} from '@/app/admin/(with-sidebar)/recruitment/_components/manage-mail/ManageMail';
-import {MailSelect} from '@/app/admin/(with-sidebar)/results/_components/MailSelect';
+import {MailSelect} from '@/app/admin/(with-sidebar)/results/_components/mail-manage/MailSelect';
 
 export const ManageResultMail = () => {
   const [activeTab, setActiveTab] = useState('합격자 메일');
