+++ src/app/admin/(with-sidebar)/results/page.tsx
@@ -1,6 +1,6 @@
 import {SuspenseWrapper} from '@/components/wrappers/SuspenseWrapper';
-import {ManageResult} from '@/app/admin/(with-sidebar)/results/_components/ManageResult';
-import {ManageResultMail} from '@/app/admin/(with-sidebar)/results/_components/ManageResultMail';
+import {ManageResult} from '@/app/admin/(with-sidebar)/results/_components/mail-manage/ManageResult';
+import {ManageResultMail} from '@/app/admin/(with-sidebar)/results/_components/mail-manage/ManageResultMail';
 
 export default function AdminResultsPage() {
   return (
