+++ src/hooks/useManageMail.ts
@@ -2,7 +2,7 @@
 
 import {useState} from 'react';
 import {useRecruitmentStore} from '@/store/useRecruitmentStore';
-import {MAIL_DATA_MAP} from '@/schemas/admin-result-type';
+import {MAIL_DATA_MAP} from '@/constants/admin/admin-result';
 
 export const useManageMail = (
   mailType: string,
